Moved srv1 and srv2 packages out of pymate, fixed imports, added dependencies to setup.py
