+++ pymate/matenet/mx.py
@@ -8,7 +8,7 @@ __author__ = 'Jared'
 
 from struct import Struct
 from pymate.value import Value
-from matenet import MateDevice, MateNET
+from . import MateDevice, MateNET
 
 
 class MXStatusPacket(object):
