+++ srv1/requirements.txt
@@ -1,0 +1,2 @@
+matplotlib==2.0.2
+numpy
