+++ srv2/bachnet.fcgi
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 from flup.server.fcgi import WSGIServer
-from receiver import app
+from .receiver import app
 
 if __name__ == '__main__':
     WSGIServer(app).run()
